baihe_api/requirements.txt
Django==1.10.4
djangorestframework==3.5.3
Pillow==4.0.0
psycopg2==2.6.2
django-ckeditor==5.2.1
XlsxWriter==0.9.6
